+++ src/main.rs
@@ -1,12 +1,10 @@
-use anyhow::{Context, Result};
+mod cli;
+mod package_manager;
+mod scripts;
+
+use anyhow::Result;
 use clap::Parser;
-use serde_json::Value;
-use skim::prelude::*;
-use std::{
-    fs,
-    path::Path,
-    process::{Command, Stdio},
-};
+use std::path::Path;
 
 #[derive(Parser, Debug)]
 #[command(author, version, about, long_about = None)]
@@ -14,23 +12,6 @@ struct Args {
     /// Path to the project directory (defaults to current directory)
     #[arg(short, long)]
     path: Option<String>,
-}
-
-/// Represents the supported package managers
-#[derive(Debug, Clone, Copy)]
-enum PackageManager {
-    Npm,
-    Pnpm,
-}
-
-impl PackageManager {
-    /// Returns the command string for this package manager
-    fn command(&self) -> &'static str {
-        match self {
-            PackageManager::Npm => "npm",
-            PackageManager::Pnpm => "pnpm",
-        }
-    }
 }
 
 /// Main function that orchestrates the script selection and execution process.
@@ -42,11 +23,11 @@ impl PackageManager {
 /// - Failed to read or parse package.json
 /// - Failed to execute the selected script
 fn main() -> Result<()> {
-    let args: Args = Args::parse();
+    let args = cli::parse_args();
 
     // Change to the specified directory or use current directory
     if let Some(path) = &args.path {
-        change_directory(path)?;
+        cli::change_directory(path)?;
     }
 
     // Check if package.json exists
@@ -56,8 +37,8 @@ fn main() -> Result<()> {
     }
 
     // Determine package manager and get scripts
-    let package_manager = determine_package_manager();
-    let scripts = get_scripts_from_package_json()?;
+    let package_manager = package_manager::determine_package_manager();
+    let scripts = scripts::get_scripts_from_package_json()?;
 
     if scripts.is_empty() {
         eprintln!("Error: There are no scripts in package.json");
@@ -65,140 +46,13 @@ fn main() -> Result<()> {
     }
 
     // Select script using fuzzy finder
-    let script_name = select_script(&scripts)?;
+    let script_name = scripts::select_script(&scripts)?;
     if script_name.is_empty() {
         return Ok(());
     }
 
     // Run the selected script
-    run_script(package_manager, &script_name)?;
-
-    Ok(())
-}
-
-/// Changes the current working directory to the specified path.
-///
-/// # Arguments
-///
-/// * `path` - The path to change to
-///
-/// # Errors
-///
-/// Returns an error if the directory change fails
-fn change_directory(path: &str) -> Result<()> {
-    std::env::set_current_dir(path)
-        .with_context(|| format!("Failed to change directory to {}", path))
-}
-
-/// Determines which package manager to use based on lock files.
-///
-/// Returns `PackageManager::Pnpm` if pnpm-lock.yaml exists, otherwise `PackageManager::Npm`.
-fn determine_package_manager() -> PackageManager {
-    if Path::new("pnpm-lock.yaml").exists() {
-        PackageManager::Pnpm
-    } else {
-        PackageManager::Npm
-    }
-}
-
-/// Extracts scripts from package.json file.
-///
-/// # Returns
-///
-/// A vector of strings in the format "script_name = script_command"
-///
-/// # Errors
-///
-/// Returns an error if:
-/// - package.json cannot be read
-/// - package.json cannot be parsed as valid JSON
-fn get_scripts_from_package_json() -> Result<Vec<String>> {
-    // Read package.json
-    let package_json = fs::read_to_string("package.json").context("Failed to read package.json")?;
-
-    // Parse package.json
-    let package_data: Value =
-        serde_json::from_str(&package_json).context("Failed to parse package.json")?;
-
-    // Extract scripts
-    match &package_data["scripts"] {
-        Value::Object(scripts_obj) => {
-            if scripts_obj.is_empty() {
-                return Ok(Vec::new());
-            }
-
-            Ok(scripts_obj
-                .iter()
-                .map(|(key, value)| format!("{} = {}", key, value.as_str().unwrap_or("")))
-                .collect())
-        }
-        _ => Ok(Vec::new()),
-    }
-}
-
-/// Presents a fuzzy finder interface for script selection.
-///
-/// # Arguments
-///
-/// * `scripts` - A slice of strings containing script names and commands
-///
-/// # Returns
-///
-/// The selected script name or an empty string if no selection was made
-///
-/// # Errors
-///
-/// Returns an error if the fuzzy finder fails to run
-fn select_script(scripts: &[String]) -> Result<String> {
-    // Use skim for fuzzy selection
-    let options = SkimOptionsBuilder::default()
-        .height(Some("50%"))
-        .reverse(true)
-        .build()
-        .context("Failed to build skim options")?;
-
-    let item_reader = SkimItemReader::default();
-    let items = item_reader.of_bufread(std::io::Cursor::new(scripts.join("\n")));
-
-    let selected_items = Skim::run_with(&options, Some(items))
-        .map(|out| out.selected_items)
-        .unwrap_or_default();
-
-    if selected_items.is_empty() {
-        return Ok(String::new());
-    }
-
-    // Extract the script name from the selected item
-    let selected = selected_items[0].output();
-    let script_name = selected.split('=').next().unwrap_or("").trim();
-
-    Ok(script_name.to_string())
-}
-
-/// Runs the selected script with the appropriate package manager.
-///
-/// # Arguments
-///
-/// * `package_manager` - The package manager to use
-/// * `script_name` - The name of the script to run
-///
-/// # Errors
-///
-/// Returns an error if the script execution fails to start
-fn run_script(package_manager: PackageManager, script_name: &str) -> Result<()> {
-    let cmd = package_manager.command();
-    println!("Running: {} run {}", cmd, script_name);
-
-    let status = Command::new(cmd)
-        .args(["run", script_name])
-        .stdout(Stdio::inherit())
-        .stderr(Stdio::inherit())
-        .status()
-        .with_context(|| format!("Failed to execute {} run {}", cmd, script_name))?;
-
-    if !status.success() {
-        eprintln!("Command failed with exit code: {:?}", status.code());
-    }
+    package_manager::run_script(package_manager, &script_name)?;
 
     Ok(())
 }
